+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * :bug:`-` Some specific 0x settler swaps in Optimism will now be properly decoded.
 * :bug:`-` When exporting history as CSV the events will now properly appear sorted by timestamp.
 * :bug:`10087` Users will now be able to use the electron app wallet bridge when using Windows.
+* :bug:`-` Token detection will no longer fail when encountering certain NFTs with improper implementations.
 
 * :release:`1.39.0 <2025-06-04>`
 * :feature:`-` DigixDAO DGD refunds will now be properly decoded.